g2g crypto module
