+++ src/pvHelpers/crypto/pv_key_factory.py
@@ -1,12 +1,21 @@
-from .key_protocols import ASYMM_KEY_PROTOCOL_VERSION, SYMM_KEY_PROTOCOL_VERSION, USER_KEY_PROTOCOL_VERSION, \
-    SIGN_KEY_PROTOCOL_VERSION
-from .user_key import UserKeyV0, PublicUserKeyV0, UserKeyV1, PublicUserKeyV1
-from .symm_key import SymmKeyV0, SymmKeyV1
-from .asymm_key import AsymmKeyV0, AsymmKeyV2, AsymmKeyV3, PublicKeyV3, PublicKeyV2, AsymmKeyV1, PublicKeyV1
-from .sign_key import SignKeyV3, SignKeyV1, SignKeyV0, VerifyKeyV1, VerifyKeyV0, VerifyKeyV3
-from .utils import CryptoException, g_log, UserKeyBuffer, ProtobufErrors, KeyBuffer, PublicUserKeyBuffer, b64dec, \
-    utf8Decode, jloads, EC_SECRET_LENGTH
-from pvHelpers import EncodingException
+from pvHelpers.crypto.asymm_key import (AsymmKeyV0, AsymmKeyV1, AsymmKeyV2,
+                                        AsymmKeyV3, PublicKeyV1, PublicKeyV2,
+                                        PublicKeyV3)
+from pvHelpers.crypto.sign_key import (SignKeyV0, SignKeyV1, SignKeyV3,
+                                       VerifyKeyV0, VerifyKeyV1, VerifyKeyV3)
+from pvHelpers.crypto.symm_key import SymmKeyV0, SymmKeyV1
+from pvHelpers.crypto.user_key import (PublicUserKeyV0, PublicUserKeyV1,
+                                       UserKeyV0, UserKeyV1)
+from pvHelpers.logger import g_log
+from pvHelpers.utils import EncodingException, b64dec
+
+from .key_protocols import (ASYMM_KEY_PROTOCOL_VERSION,
+                            SIGN_KEY_PROTOCOL_VERSION,
+                            SYMM_KEY_PROTOCOL_VERSION,
+                            USER_KEY_PROTOCOL_VERSION)
+from .utils import (EC_SECRET_LENGTH, CryptoException, KeyBuffer,
+                    ProtobufErrors, PublicUserKeyBuffer, UserKeyBuffer)
+
 
 class PVKeyFactory(object):
     @staticmethod
