+++ src/pvHelpers/crypto/user_key/__init__.py
@@ -1,3 +1,3 @@
-from .user_key_base import *
-from .user_key_v0 import *
-from .user_key_v1 import *
+from .user_key_base import PublicUserKeyBase, UserKeyBase
+from .user_key_v0 import PublicUserKeyV0, UserKeyV0
+from .user_key_v1 import PublicUserKeyV1, UserKeyV1
