+++ src/pvHelpers/mail/__init__.py
@@ -1,4 +1,4 @@
-from .email import *
-from .prepared_message import *
+from .email.email_base import EmailBase
+from .prepared_message.prepared_message_base import PreparedMessageBase
 from .email_factory import EmailFactory
 from .prepared_message_factory import PreparedMessageFactory
