+++ src/pvHelpers/mail/email/__init__.py
@@ -1,9 +1,9 @@
-from .content import Content
 from .attachment import Attachment, AttachmentMetadata, AttachmentType
-from .server_attributes import ServerAttributes
-from .email_helpers import EmailException, EmailHelpers, PROTOCOL_VERSION, DUMMY_CONTENT_TYPE, DUMMY_DISPOSITION
+from .content import Content
 from .email_base import EmailBase
+from .email_helpers import EmailException, EmailHelpers
 from .email_v1 import EmailV1
 from .email_v2 import EmailV2
 from .email_v3 import EmailV3
 from .email_v4 import EmailV4
+from .server_attributes import ServerAttributes
