+++ src/pvHelpers/user/__init__.py
@@ -1,6 +1,6 @@
-from .device import *
-from .local_device import *
-from .user import *
-from .local_user import *
+from .device import Device, DeviceStatus
+from .local_device import LocalDevice
+from .local_user import LocalUser
 from .luser_info import *
+from .user import User
 from .user_db_node import UserDBNode
